Removing unused screenSize key, renaming Context -> Toolbox (it contains more than just context-stuff). Applying some simplification suggestions from IntelliJ in Pagination.tsx.
