+++ src/components/Table/index.ts
@@ -4,7 +4,7 @@ export { TableRow } from './TableRow';
 export { TableBody } from './TableBody';
 export { TableCell } from './TableCell';
 export { TableFooter } from './TableFooter';
-export type { ChangeProps, SortProps } from './Context';
+export type { ChangeProps, SortProps } from './Toolbox';
 export { SortDirection } from './TableCell';
 export type { RowData } from './TableRow';
-export { ScreenSize } from './Context';
+export { ScreenSize } from './Toolbox';
